offline stuff

+++ src/services/ApiService.js
@@ -5,9 +5,25 @@ import AuthService from "./AuthService";
 class ApiService {
   constructor() {
     axios.defaults.baseURL = ConfigService.get("api_base_url");
+    this.store = null;
+  }
+
+  init(store) {
+    this.store = store;
+  }
+
+  isOffline() {
+    const state = this.store.getState();
+    if (state.core.meta.connected) {
+      return false;
+    } else {
+      alert("You are offline, Toodu is in read-only mode.");
+      return true;
+    }
   }
 
   post(url, data) {
+    if (this.isOffline()) return;
     return axios
       .post(url, data, { headers: AuthService.auth })
       .then(this.setAuth);
@@ -20,12 +36,14 @@ class ApiService {
   }
 
   put(url, data) {
+    if (this.isOffline()) return;
     return axios
       .put(url, data, { headers: AuthService.auth })
       .then(this.setAuth);
   }
 
   delete(url) {
+    if (this.isOffline()) return;
     return axios.delete(url, { headers: AuthService.auth }).then(this.setAuth);
   }
